:truck: moved identification to shared
