+++ minimark/src/rmrk/index.ts
@@ -1,6 +1,6 @@
 export * from './shared/constants'
 export * from './v1/create' // TODO: Deprecate
 export * from './shared/helpers'
-export * from './identification'
+export * from './shared/identification'
 export * from './types'
 export * from './unwrap'
