+++ minimark/src/rmrk/shared/identification.ts
@@ -1,7 +1,6 @@
 import { nanoid } from 'nanoid'
-import { addressToHex } from '../utils/hex'
-import { toUpperCase, upperTrim } from '../utils/string'
-import { CreatedNFT } from './types'
+import { addressToHex } from '../../utils/hex'
+import { isEmptyString, toUpperCase, upperTrim } from '../../utils/string'
 
 export const toSerialNumber = (index: number, offset = 0, plusOne = true): string => {
   return String(index + offset + Number(plusOne)).padStart(16, '0') // TODO RMRK v2 has 8 digits
@@ -21,13 +20,12 @@ export const makeSymbol = (symbol?: string): string => {
   return !symbol ? toUpperCase(nanoid(13)) : upperTrim(symbol, true)
 }
 
-export const toNFTId = (nft: CreatedNFT, blocknumber: string | number): string => {
-  const { collection, instance, sn } = nft
-  if (!collection || !instance || !sn) {
-    throw new ReferenceError('[APP] toNFTId: invalid nft')
+export const toNFTId = (collectionId: string, symbol: string, serialNumber: string, blocknumber: string | number): string => {
+  if (isEmptyString(collectionId) || isEmptyString(symbol) || isEmptyString(serialNumber)) {
+    throw new ReferenceError('[MINIMARK] toNFTId: invalid nft')
   }
 
-  return `${blocknumber}-${collection}-${instance}-${sn}`
+  return `${blocknumber}-${collectionId}-${symbol}-${serialNumber}`
 }
 
 export const findUniqueSymbol = (symbol: string | undefined, usedSymbols: string[]): string => {
